+++ db/default_song_library.sql
@@ -1,0 +1,3 @@
+INSERT INTO bot_default_song_library (id, artist, song_title, album, length_in_seconds) VALUES
+(1, "test1", "test1", "test1", 5),
+(2, "test2", "test2", "test2", 5);